# lambda/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the Biclomap back-end server",
        "title": "Biclomap REST API",
        "contact": {
            "name": "API Support",
            "url": "http://dev.biclomap.com/support",
            "email": "api-support@biclomap.com"
        },
        "license": {
            "name": "GPL v3.0"
        },
        "version": "1"
    },
    "host": "dev.biclomap.com",
    "basePath": "/api",
    "paths": {
        "/login/email": {
            "post": {
                "description": "Users having succesffuly signed-up and then succesfully having\nconfirmed their e-mai will be able to effectively login into\nthe application using this method call. The system will create\na JWT token. Any subsequent API call will need to provide this\ntoken in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Perform email-based login and return JWT token",
                "parameters": [
                    {
                        "description": "Email and password structure",
                        "name": "email_login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.EMAIL_LOGIN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/login.LOGIN_INFORMATION"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/login/email/confirm": {
            "post": {
                "description": "This method should be called when the user clicks the link\nprovided in the e-mail sent by the email signup call. This\nmethod will effectively activate the user account and, if\nsuccessful, the user will be able to login using the e-mail\nand password provided upon calling the email signup method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribe"
                ],
                "summary": "Confirms an e-mail",
                "parameters": [
                    {
                        "description": "email confirmation structure",
                        "name": "email_confirm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.EMAIL_CONFIRM"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/login/email/signup": {
            "post": {
                "description": "This method starts the new user registration process. It adds\na new record to the database. The password is stored in hashed form. Then\nit automatically sends an e-mail to the email address given in the parameter\nThat e-mail contains a verification token, that's also computed here.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscribe"
                ],
                "summary": "Signup new user by their e-mail",
                "parameters": [
                    {
                        "description": "Email signup",
                        "name": "email_signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.EMAIL_SIGNUP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login/facebook": {
            "post": {
                "description": "This is called after user successfully loged-in with Facebook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Open session for facebook-authenticated user",
                "parameters": [
                    {
                        "description": "login structure",
                        "name": "fb_login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.FB_LOGIN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/login.LOGIN_INFORMATION"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Get the status of the API",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "HealthCheck"
                ],
                "summary": "Show the status of the API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "login.EMAIL_CONFIRM": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "login.EMAIL_LOGIN": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "login.EMAIL_SIGNUP": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "login.FB_LOGIN": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "login.LOGIN_INFORMATION": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}